Fix hidden state flickering when disabled
The enabled state was the default state to make sure that we wouldn't have any personal detail could flickering in the screen. But this was making the normal navigation in Github weird because the hidden state would always briefly appear.

We're now adding a temporary blank state to the sensitive data while the page is loading, which solves the problem. It's not perfect because the blank state will appear regardless of the extension being active or not, but it's a lot less annoying.